pyproject.toml
[project]
name = "real-estate-classifier"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "matplotlib>=3.10.8",
    "pillow>=11.3.0",
    "python-multipart>=0.0.26",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
    "timm>=1.0.26",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.44.0",
    "wandb>=0.26.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.3",
]
